#ue5 - Fix truncation warnings in modules: DeviceProfileServices, FieldSystemEngine, FontEditor, FunctionalTesting. Includes some related engine fixes as well.
#jira UE-160832
#rb Andrew.Davidson, Dave.Jones2
#preflight 632cbdc17b582f58ab6315ad

[CL 22164453 by zak middleton in ue5-main branch]

+++ Engine/Source/Developer/FunctionalTesting/Classes/FunctionalTest.h
@@ -35,7 +35,7 @@ struct FStatsData
 	FStatsData() :NumFrames(0), SumTimeSeconds(0.0f){}
 
 	uint32 NumFrames;
-	uint32 SumTimeSeconds;
+	float SumTimeSeconds;
 	FStatisticalFloat FrameTimeTracker;
 	FStatisticalFloat GameThreadTimeTracker;
 	FStatisticalFloat RenderThreadTimeTracker;
